requirements.txt
fastapi==0.121.3
fitz==0.0.1.dev2
motor==3.7.1
openai==2.8.1
Pillow==12.0.0
pydantic==2.12.4
pymongo==4.15.4
pypdf==6.3.0
python-dotenv==1.2.1
